install package, merge
