+++ src/pages/CartPage.js
@@ -1,0 +1,7 @@
+import React from 'react';
+import styled from 'styled-components';
+
+const CartPage = () => {
+   return <>Cart PAGE</>;
+};
+export default CartPage;
