+++ src/pages/ReviewWritePage.js
@@ -1,0 +1,7 @@
+import React from 'react';
+import styled from 'styled-components';
+
+const ReviewWritePage = () => {
+   return <>ReviewWrite PAGE</>;
+};
+export default ReviewWritePage;
