+++ src/pages/SearchPage.js
@@ -1,0 +1,7 @@
+import React from 'react';
+import styled from 'styled-components';
+
+const SearchPage = () => {
+   return <>Search PAGE</>;
+};
+export default SearchPage;
